requirements.txt
# python >= 2.7
ansible~=3.0.0
flask~=1.1.2
setuptools~=54.0.0
pyyaml >= 5.4.1
pandas
pymongo
dnspython
virtualenvwrapper
openpyxl
djangorestframework
django-cors-headers
prometheus-api-client
django-jsonforms
Django==2.0.3
gunicorn==19.7.1
Pillow==4.3.0
requests==2.18.4
six==1.9.0
stream-django==1.4.0
whitenoise==3.3.1
autopep8==1.3.2
django-allauth==0.33.0
rollbar==0.13.13
dj-database-url==0.5.0
djongo
zenml